Modify lfu cache to shrink memory usage
In the original LFU cache implementation, the memory usage fluctuates
with changes in the THRESHOLD setting. Consequently, when we increase
the THRESHOLD, the memory usage increases significantly. To address this
issue, I have modified the implementation of LFU cache lists. Now, memory
is allocated for these lists only when a cache entry is appended to them.

The original memory usage for cache lists was THRESHOLD (32678) * sizeof(struct list_head).
The proposed implementation reduces it to cache_size (1024) * sizeof(struct list_head).