Implement tier-2 JIT compiler to accelerate ISS
The T2C, designed for flexibility, efficiently transforms custom IR
within a chained block into LLVM IR. After the translation, the built
LLVM IR is offloaded to the LLVM backend, where it undergoes
optimization through several selected LLVM passes. Subsequently, the
optimized LLVM IR is passed to the LLVM execution engine, which compiles
the optimized LLVM IR and returns a function pointer to the generated
machine code. Because our design for T2C opts for a function pointer of
the generated machine code instead of storing it in the code cache, we
need not account for the binary format discrepancies among different
compilers. This process allows for the possibility of substituting LLVM
with alternative optimization frameworks if required.